Le premier commit du frontEnd

+++ src/app/components/Courriers/listeCourrier/liste-courrier/liste-courrier.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {CourrierService} from "../../../../services/Courriers/courrier.service";
 import {Router} from "@angular/router";
+import {AlertService} from "../../../../services/Alerts/alert.service";
 
 @Component({
   selector: 'app-liste-courrier',
@@ -9,7 +10,7 @@ import {Router} from "@angular/router";
 })
 export class ListeCourrierComponent implements OnInit {
   listCourrier : any = [];
-  constructor(private courrierService : CourrierService, private router : Router) { }
+  constructor(private courrierService : CourrierService, private router : Router, private alertService : AlertService) { }
 
   ngOnInit(): void {
     this.getAllCourriersFromService();
@@ -29,10 +30,17 @@ export class ListeCourrierComponent implements OnInit {
   }
 
   supprimerCourrier(id : any, i : any){
-    this.courrierService.deleteCourrierFromServer(id).subscribe((resultat)=>{
-      this.listCourrier.splice(i, 1)
-    }, (erreur)=>{
-      console.log(erreur)
+    this.alertService.emettreConfirmationAlert("Voulez vous supprimer ce courrier", "ce courrier").then((resultat: { isConfirmed: any; })=>{
+      if(resultat.isConfirmed){
+        this.courrierService.deleteCourrierFromServer(id).subscribe((resultat)=>{
+          this.listCourrier.splice(i, 1)
+          this.alertService.emettreUnToast("Suppression de courrier !", 'success');
+          this.router.navigate(['/home', 'courrier', 'liste'])
+        }, (erreur)=>{
+          this.alertService.emettreUnToast("Suppression de courrier !", 'error');
+        })
+      }
     })
+
   }
 }
