+++ src/app/components/Roles/listeRole/liste-role/liste-role.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {Router} from "@angular/router";
 import {RoleService} from "../../../../services/Roles/role.service";
+import {AlertService} from "../../../../services/Alerts/alert.service";
 
 @Component({
   selector: 'app-liste-role',
@@ -11,7 +12,7 @@ export class ListeRoleComponent implements OnInit {
 
   listeRole : any = [];
 
-  constructor(private roleService : RoleService, private router : Router) { }
+  constructor(private roleService : RoleService, private router : Router, private alertService : AlertService) { }
 
   ngOnInit(): void {
     this.getAllRolesFromRoleServer();
@@ -30,11 +31,16 @@ export class ListeRoleComponent implements OnInit {
   }
 
   supprimerRole(_id: any, i: number) {
-    this.roleService.deleteRoleFromServer(_id).subscribe((resultat)=>{
-      this.listeRole.splice(i, 1)
-    }, (erreur)=>{
-      console.log(erreur)
+    this.alertService.emettreConfirmationAlert("Voulez vous supprimer ce rôle","ce rôle").then((resultat: { isConfirmed: any; })=>{
+      if(resultat.isConfirmed){
+        this.roleService.deleteRoleFromServer(_id).subscribe((resultat)=>{
+          this.listeRole.splice(i, 1)
+          this.router.navigate(['/home', 'role', 'liste'])
+          this.alertService.emettreUnToast("La suppression du rôle a été effectuée avec succès !", 'success');
+        }, (erreur)=>{
+          this.alertService.emettreUnToast("Echec de la suppression du rôle !", 'error');
+        })
+      }
     })
-
   }
 }
